[feat] response region / theater dto 수정
