--- app/src/main/java/com/sopt/lottecinemaaos/data/entity/Theater.kt
@@ -1,12 +1,0 @@
-package com.sopt.lottecinemaaos.data.entity
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Theater(
-    @SerialName("theater_id")
-    val theaterId: Int,
-    @SerialName("theater_name")
-    val theaterName: String
-)
