+++ app/src/main/java/com/sopt/lottecinemaaos/data/model/response/ResponseRegionDto.kt
@@ -1,6 +1,6 @@
 package com.sopt.lottecinemaaos.data.model.response
 
-import com.sopt.lottecinemaaos.data.entity.Region
+import com.sopt.lottecinemaaos.domain.model.Region
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -12,5 +12,22 @@ data class ResponseRegionDto(
     val message: String,
     @SerialName("data")
     val data: List<Region>
+) {
+    @Serializable
+    data class Region(
+        @SerialName("region_id")
+        val regionId: Int,
+        @SerialName("region_name")
+        val regionName: String,
+        @SerialName("theater_count")
+        val theaterCount: Int
+    )
 
-)
+    fun toRegion() = data.map { region ->
+        Region(
+            id = region.regionId,
+            name = region.regionName,
+            count = region.theaterCount
+        )
+    }
+}
