+++ app/src/main/java/com/sopt/lottecinemaaos/domain/model/Region.kt
@@ -1,0 +1,8 @@
+package com.sopt.lottecinemaaos.domain.model
+
+data class Region(
+    val id: Int,
+    val name: String,
+    val count: Int
+
+)
