+++ app/src/main/java/com/sopt/lottecinemaaos/domain/model/Theater.kt
@@ -1,0 +1,6 @@
+package com.sopt.lottecinemaaos.domain.model
+
+data class Theater(
+    val id: Int,
+    val name: String
+)
